Move Java SDK adapter to a new naming strategy
