+++ aws/sdkv2-retries/build.gradle.kts
@@ -4,8 +4,8 @@ plugins {
 
 description = "This module provides an adapter that allows Smithy Java to use retry strategies from the AWS SDK for Java V2"
 
-extra["displayName"] = "Smithy :: Java :: Retries :: AWS SDK for Java Adapter"
-extra["moduleName"] = "software.amazon.smithy.java.retries.sdkadapter"
+extra["displayName"] = "Smithy :: Java :: AWS :: AWS SDK v2 :: Retries adapter"
+extra["moduleName"] = "software.amazon.smithy.java.aws.sdkv2.retries"
 
 dependencies {
     api(project(":retries-api"))
