+++ aws/sdkv2-retries/src/main/java/software/amazon/smithy/java/aws/sdkv2/retries/SdkRetryToken.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.java.retries.sdkadapter;
+package software.amazon.smithy.java.aws.sdkv2.retries;
 
 import software.amazon.smithy.java.retries.api.RetryToken;
