+++ aws/sdkv2-retries/src/test/java/software/amazon/smithy/java/aws/sdkv2/retries/SdkRetryStrategyTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.java.retries.sdkadapter;
+package software.amazon.smithy.java.aws.sdkv2.retries;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
